pyproject.toml
[tool.poetry]
name = "plwn-graph-builder"
version = "1.0"
description = "Mapping plWordNet from the database structure into Lexical Units or Synsets graph."
authors = ["Łukasz Kopociński <lkopocinski@gmail.com>"]
readme = "README.md"
repository = "https://github.com/lkopocinski/plwn-graph-builder"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.0"
PyYAML = "^5.3"
sqlalchemy = "^1.3"

[tool.poetry.dev-dependencies]
flake8-bugbear = "^19.8"
isort = "^4.3.20"
black = { version = "^19.10b0", python = "^3.8" }
pre-commit = "^2.3.0"
ipython = "^7.7"

[tool.isort]
line_length = 90
multi_line_output = 3
lines_between_types = 1
lines_after_imports = 2

[tool.black]
line-length = 88
target_version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/

)
'''

[tool.poetry.scripts]
plwn_builder = 'plwn_graph_builder.main:main'

[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"
